refactor: split async/sync server cursor in two modules
